requirements.txt
numpy
sounddevice
keyboard
scipy
assemblyai
pyttsx3
python-dotenv
google-genai
pyserial
requests